Script tests the ability to recognize faces
Script loads in a folder of images then resizes them to be similar then
randomly inverts a selection of them and displays to user.  At end of
display task, user is asked if he/she recognizes the face shown.
Experiment is performed in order to determine the effects of an unusual
viewing angle affects memory either positively or negatively.