Experiment that displays color words in the wrong color as the text
Script demonstrates a classic psychology experiment whereupon a color
word is shown in the wrong color.  Script takes in user input as to
whether or not the word color matches the word displayed. Word
congruency and incongruency is randomly generated and user input is
recorded and data is calculated at the end.